fix: Stop the container via SIGINT
This is a workaround for the `special` signal handling of ros2 launch.
To stop a process docker sends a SIGTERM on default.
The ros2 launch wrapper however ignores SIGTERM (only SIGINT is
considered).
Docker then sends a SIGKILL after a timeout.
As a result, the shutdown times of the container are at least as long as the
timeout.
This change makes docker send SIGINT instead of SIGTERM, resulting in
proper fast shutdown.
See https://github.com/ros2/launch/issues/666 for details.