test: Assert that a service call actually reaches the OPC UA server
Previously it was only asserted that the ROS service call was registered
by the bridge but not if it was actually passed through to the OPC UA
server.
With this change, the assertion channel of the OPC UA test server is
levaraged in order to cover more of the code path under test.